backend/requirements.txt
accelerate==1.12.0
aiofiles==24.1.0
aistudio-sdk==0.3.8
albucore==0.0.24
albumentations==2.0.6
alembic==1.18.4
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
asyncpg==0.31.0
bce-python-sdk==0.9.59
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.2.0
cloudevents==1.12.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
Cython==3.1.0
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
duckduckgo_search==8.1.1
easyocr==1.7.2
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.9.23
fonttools==4.58.0
fsspec==2025.3.2
functions-framework==3.8.3
future==1.0.0
googlesearch-python==1.3.0
gradio==5.29.0
gradio_client==1.10.0
greenlet==3.4.0
groovy==0.1.2
groq==0.24.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
lmdb==1.6.2
lxml==5.4.0
Mako==1.3.11
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
MIDIUtil==1.2.1
ml_dtypes==0.5.3
modelscope==1.34.0
mpmath==1.3.0
munch==4.0.0
networkx==3.1
ninja==1.11.1.4
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
pretrainedmodels==0.7.4
prettytable==3.17.0
primp==0.15.0
protobuf==6.30.2
psutil==7.1.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pypdfium2==5.3.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2025.11.3
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.19.1
ruff==0.11.9
safehttpx==0.1.6
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
semantic-version==2.10.0
setuptools==78.1.0
shapely==2.1.0
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.49
starlette==0.46.2
stringzilla==3.12.5
sympy==1.14.0
termcolor==3.1.0
tifffile==2025.5.10
tokenizers==0.22.1
tomlkit==0.13.2
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.57.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.11.0
urllib3==2.3.0
uvicorn==0.34.2
vino==0.0.2
watchdog==6.0.0
wcwidth==0.2.14
websockets==15.0.1
PyJWT